refactor(run): move run functions from main.py to own module
